Add lzma/xz compression support
This patch adds support for the lzma/xz2 compression format using the `xz2`
crate.  The implementation is hidden behing the 'lzma' feature.  The feature is
not enabled by default because the `xz2` crate (or more accurately the
`libz-sys` crate `xz2` depends on) requires rust version 1.26+ (it uses the
`fs::read_to_string` function from the standard library).